Major refactor.
Added folding to the init.vim, with it on by default.
Lots of new settings, and totally re-orgnized the file by area.
It now has a nice index when you open it up.

Now nvim saves the last cursor position in a file.
A bit of Go Lang support.

Python support.

Removed some unused plugins.